Update neovim configuration and hyprland configuration
Full removal of lspsaga from the neovim configuration. It was
conflicting too badly with the other keybinds that I had and was messing
with the way the window bar is displayed. Hyprland configuration and
waybar are adjusted so that the main modules use the theme colors as an
outline instead of background.