fix(nvim): Finalize revisions to keymaps and plugin functionality
In a previous commit c92530a, the first underpinnings of revision to the
keymappings was staged. This commit finalizes the changes made in that
commit and provides additional debugging as to the strange behavior
encountered behind the "n" and "v" key remappings and the fact that our
which-key descriptions were not being correctly displayed in the popup
cheatsheet.

Coincidentally, this commit also helps to remedy issues related to
consistent themeing of UI elements using the "shadow" border style.