fix(nvim): Update core modules further to improve UI unification.
We are getting very close to having this to a "completed" point.
Admittedly it may never reach that stage explicitly, but from the
abstract, more general perspective, the core plugin components and
registration seems to be working correctly.

We still need to fix the UI further such that the completion menu
displays with the correct highlighting.