debug(lualine): Fix an issue of plugin timing with line and dash
The dashboard plugin used was having some trouble with messing up the display
settings for the lualine plugin. We also can't load the dashboard on the
VeryLazy event, as it doesn't function correctly. But we also need to load
lualine at the same time as the dashboard, otherwise the compatibility goes
haywire, and for certain files the top tabline doesn't work.