feat(lsp): Prepare for the use of inbuilt inlay-hints in neovim.
These aren't out yet on the main build of neovim but they do exist on the
nightly build if you build it from source, or so I hear. But we may just end up
waiting until they have successfully implemented these on the main branch and
the version hits the repositories.