feat(ascii): update ascii insertion methods to fix helplines, add dash header
This commit handles some adjustments made for the purposes of any banner
creation using figlet or which have been included in a secondary addition to the
standard splash-screen, which is much more stripped back/minimal. Adds the
fsplash plugin which is supposed to splash entry on the main dashboard with the
nightowl banner, currently a bit broken it seems.