feat(mini): remove mini.sessions dependency, keep the spec for other reasons
mini.sessions is a great plugin but I don't find myself using it all too often
to manage my workspace. Generally I prefer to start fresh. It would be ideal to
have a more populated project selection area though. edgy.nvim should be able to
help in that regard in the future. so would neorg.