observe: note that present-imperative style is specifically and important.
These updates mainly seek to reorganize some of the file management. The aim
here is that ideally our "actions" within the module correspond roughly to
things that can be done in a git setup with a more conventional structure. Also
does a shift-around of some unused keymappings to make the ergonomics of using
nightowl.nvim slightly better.