feat(alpha): move environment spec for dashboard to a different file
The plugin itself is in a minor state of unfinished implementation, but we would
like to use Veil as the primary dashboard because it is more customizable and
feels more like how I want a dashboard to be. This commit moves some variables
into the de-facto spot for user configuration, e.g. the environment directory.