update(lsp): adjust plugin load order specifications at lsp level
This is in search of a fix to the elusive problem which is the notification on
startup that neoconf should be run before any lspconfig setup. Previously, it
seems as though there was some kind of circular dependency introduced during the
core lsp initialization due to lsp-zero being included superfluously in
dependency fields of certain lsp plugins.