fix(keys): move core keybindings to some slightly different configurations
Playing around with some new placements of the keymappings in order to better
facilitate the difficult to balance combination of semantic meaning through
key-mappings, which-key functionality, and vim default functionality. For
instance, in an ideal scenario, `q` wouldn't do anything related to recording
macros, so that it can be used for window management (e.g. quitting, hence q).