fix(lazy): adjust lazy specifications for behavioral type plugins
Mainly this removes any opts fields that I have found which are completely empty
and don't have a strange specification on the config field for each plugin
specification. Also, removes some events or adjusts the events on which plugins
get loaded, so as to not have everything loading if it is not necessary on
VeryLazy.