fix(ui): update some window settings; looking for source of error
Occasionally, it seems as though opening FZF will cause a change to the buftype
available on the main window. What occurs is that the `q` key causes a quit
event during normal mode, after FZF has been opened. Unfortunately, this could
be caused by a great many things, and could be in an entirely unrelated plugin.
Hence we have to test, starting with places that seemed the most damning.

Still haven't found it yet.