feat(windows): remove focus.nvim and replace with windows.nvim
The rationale here was that focus seemed to be causing strange, unresolvable
errors at various points or after various actions. So, we needed to get rid of
that plugin, but the functionality it provides is rather nice. Luckily there are
a few other plugins that can accomplish that, we choose windows.nvim here