feat(cmp): introduce the ability to toggle either cmp mode
The first parameter toggles the cmp activation full-stop. The second
parameter toggles only autocompletion. By default, I like this set up so
that there is no autocompletion by default, but that the plugin is
enabled. Then, in cases where cmp is helpful, the toggle can bring
autocompletion back to relevance