fix(tasks): adjust some neorg module inclusions and introduce due.nvim
The modules for Neorg are not very well-developed, yet, but they are
moving along decently well and rather quickly from what I can tell. This
is the nvim configuration that does it all though, so being able to have
multiple methods of further investigation into strange happenings is
useful, even if not immediately so.