fix(env): update environment management scripts for older individuals
These are supposed to help Mimi and Dugan stay safe online. But are also
just useful generally to have. This commit adjusts some parameters
during thawing to allow for more efficient freezing at the expense of a
slight amount of texture, probably, from being kept in the freezer. it
would be nice if we knew where one was,