fix(lsp): remove lsp-zero in default format of parliament development
While this is a very useful plugin, and I would like to use it with
LazyVim, it seems as though this is more difficult than I originally
anticipated. In particular, the nvim-lspconfig setup is handled already
by LazyVim setup, so therefore we want to avoid redefining the config
field of the corresponding Spec. LSP Zero does not work in that regard

We will still include the possibility to use LSP Zero if desired, and it
will be accessible through a LazyFlex integration.