feat(sak): Introduce the `funsak` module, a functional swiss army knife
This toolset is a collection of helpers and useful utilities that I have
created during my time working on lazy.nvim managed neovim setups. The
module seeks to provide the ability to introduce a more functional
workflow to the lazy.nvim setup for more dynamic on-the-fly
computations/configurations. And moreover, just to facilitate the
construction of a neovim configuration which can do anything, without
needing to worry too much about extra boilerplate required to do things
a certain "lazy" way.