fix(keys): Adjust keymaps so that buffer switching is smoother
The previous and default LazyVim behavior is to switch buffers when the
user presses either <S-h> or <S-l>. The new behavior prevents the switch
of buffers unintentionally, by adding a third required Ctrl modifier to
be held. This frees H and L as keys, hence we bind them to the motions
to go to line start or line end for convenience.