fix(keys): refactor the processing for the user's key mappings
We want to make sure that the key definitions are easy to follow and
make sense for the perspective of user-configuration. The previous
version, much like cmp, was a bit verbose. By moving component parts
into environment and then changing the construction of the keymap tables
that are sent off to which-key, lazy, or vim directly, we avoid having a
single long file containing all keymap definitions, instead preferring
several smaller ones with more organized semantics.