fix(lang): Fix programming language definition files for conform
Now that LazyVim has incorporated the new conform.nvim and nvim-lint
instead of the previously selected null-ls, we need to make some
adjustments to our integration of new languages. The goal here is to
make it very easy to add all component parts that must be added for a
particular language, e.g. plugins, cmp sources, formatters, and linters.