feat(lsp): dramatically update lsp configurations, achieve usable state
This update is a complete overhaul of the way that the LSP system is
configured in neovim. In particular, we closely integrate the now
required lsp-zero and set up a more separated scope for each moving
piece. Ideally, we want to keep some degree of the fluidity that comes
with not having that separated scope, but introducing the important
benefits that come with separation.