rm(null): fully deprecate and remove none-ls/null-ls
I suspect that null-ls is causing a large amount of the hangups that I seem to
be experiencing on account of the fact that it is running so many mason tools
through it. There were other ways of handling this and keeping null-ls, but its
an archived project and I prefer EFM anyhow.