feat(status): replace lualine with heirline
While lualine is a fantastic plugin, it is a bit restrictive in the way that it
specifies formatting and behavior for the statusline/tabline in neovim. This
commit replaces lualine with the equally-powerful heirline.nvim, which is
hopefully a bit faster, but mostly does not make such prescriptions about the
precise behavior or appearance of the statusline.

This commit includes component definitions, aggregations into full statusline
objects, and any helpers that are necessary to achieve these goals. It also
includes some adjustments to the behavior of incline, the rest of which follow
in the next commit.